Updated how the bot handles being teleported
Changed the bot teleporting code to be more consistent with how the base game does it. Mostly allowing base game calls to TeleportPlayer influence the rotation of the bot.

Also removed a hack added to TeleportAgentAIAndBody that was added in an attempt to fix an issue with the adamance fire exit that was caused by the old fallen bridge detection code.