Added voice command support!
- Changed voice libraries, since the old one required manual setup
- PySpeech is the new voice library
- Updated bot manager to send voice messages sent by each player so they can be "heard" by all bots.
- Updated dependencies

Something to note is that all chat commands work as voice commands. This means that saying "man the ship" is the same as typing it out in the chat. And yes, bots cannot hear your from across the map, unless both of you have walkie-talkies. I will add config options later.