Improved SearchingForScrapState's entrance safety awareness
If the bot attempts to use an entrance, but is pushed out more than 3 times, the bot will pick another entrance to the main building if possible. If there are no other available entrances, the bot will return to the ship.

Also attempted to fix an edge case with some items not being seen by bots.